+++ packages/babel-plugin-fbt/src/fbt-nodes/FbtElementNode.js
@@ -38,6 +38,7 @@ const {
   expandStringConcat,
   normalizeSpaces,
   setUniqueToken,
+  varDump,
 } = require('../FbtUtil');
 const {GENDER_ANY} = require('../translate/IntlVariations');
 const {GenderStringVariationArg} = require('./FbtArguments');
@@ -402,6 +403,47 @@ class FbtElementNode
   __toJSONForTestsOnly(): mixed {
     const ret = super.__toJSONForTestsOnly();
     return this.constructor.__compactTokenSet(ret);
+  }
+
+  assertNoOverallTokenNameCollision(
+    argsMapList: $ReadOnlyArray<StringVariationArgsMap>,
+  ): void {
+    // Importing this module only here to avoid dependency cycle
+    const FbtImplicitParamNode = require('./FbtImplicitParamNode');
+    const FbtSameParamNode = require('./FbtSameParamNode');
+
+    const tokenNameToFbtNode: {[tokenName: string]: ?AnyFbtNode} = {};
+    argsMapList.forEach(argsMap => {
+      runOnNestedChildren(this, child => {
+        if (
+          // FbtImplicitParamNode token names appear redundant but
+          // they'll be deduplicated via the token name mangling logic
+          child instanceof FbtImplicitParamNode ||
+          // FbtSameParamNode token names are allowed to be redundant by design
+          child instanceof FbtSameParamNode
+        ) {
+          return;
+        }
+        const tokenName = child.getTokenName(argsMap);
+        if (tokenName != null) {
+          const existingFbtNode = tokenNameToFbtNode[tokenName];
+          invariant(
+            existingFbtNode == null || existingFbtNode === child,
+            "There's already a token called `%s` in this %s call. " +
+              'Use %s.sameParam if you want to reuse the same token name or ' +
+              'give this token a different name.\n' +
+              'Existing FbtNode=%s\n' +
+              'Redundant FbtNode=%s',
+            tokenName,
+            this.moduleName,
+            this.moduleName,
+            varDump(existingFbtNode),
+            varDump(child),
+          );
+          tokenNameToFbtNode[tokenName] = child;
+        }
+      });
+    });
   }
 }