+++ CHANGELOG.md
@@ -149,6 +149,7 @@ List of changes for each released npm package version.
       Unreleased changes that have landed in main. Click to see more.
     </summary>
 
+    - [fix]! Prevent token name collisions among fbt constructs across all nesting levels
     - [fix] Add missing `fbjs` npm dependency
     - [feature] Introduce a script to identify new/updated strings after Auto-Param
     - [feat]! Pull out common `desc` from leaves and include `tokenAliases`